Add full example
Add additional example which makes use of all features from fault
injection.

Watch a signal which indicates a termination of the simulation.
If the design can discover an injected fault and reach a save state,
this can be considered a correct behaviour in response to a fault and
the simulation can be ended.