[verilator] Add argument parsing
Common arguments used with fault injection can be parsed directly.

This makes is necessary that the instance will live for more than one
iteration. Provide a new way to set the current iteration variable.

Now it is possible to either use parsing of the arguments and then
update the instance for each iteration or to use the `SetMode*` methods
to manually set the parameters.